step6_frontend/backend/requirements.txt
# Flask核心框架
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0

# 数据库相关
SQLAlchemy>=2.0.0,<2.1.0
alembic==1.12.1

# 数据验证和序列化
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# 密码加密
Werkzeug==2.3.7
bcrypt==4.0.1

# 文件处理
Pillow>=10.0.0

# 开发工具
python-dotenv==1.0.0

# API文档
flask-swagger-ui==4.11.1
apispec==6.3.0
apispec-webframeworks==0.5.2

# 测试
pytest==7.4.3
pytest-flask==1.3.0

# 生产环境
gunicorn==21.2.0
